backend/requirements.txt

flask
flask-cors
pymongo
PyJWT
python-dotenv
werkzeug
dnspython
requests
PyPDF2
python-docx
pdfminer.six
sentence-transformers
scikit-learn
numpy
tqdm
validators
pillow
